requirements.txt
numpy
scipy
sympy
legacy-quadpy
